+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 - Adds ability to specify a custom height and corner radius for the drawer presentation style.
 - Adds ability to grant an entitlement to anyone.
 - Adds `Superwall.shared.setIntegrationAttributes(_:)` which allows you to set attributes for third-party integrations.
+- Adds `Superwall.shared.setIntegrationAttribute(_:_:)` for setting individual integration attributes.
 - Adds `Superwall.shared.integrationAttributes` to get the attributes you've set.
 - Adds the ability to ask for an App Store review from a paywall tap action.
 - Adds a popup presentation style.